⚡ :: (#86) mypage sold, liked, bought 추가
